Main Login Panel Listener
Implemented the listener functionality to the main login panel. Future: Please change this to smaller logo. NOTE: Do not run any listeners outside of the EDT for the button. This will cause concurrency issues and kick back a NULL Pointer exception since the hashmap is being written outside of the EDT thread.